requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
requests==2.31.0
chromadb==0.4.18
google-genai==0.2.2
beautifulsoup4==4.12.2
lxml==4.9.3
pydantic==2.5.0
pydantic-settings==2.1.0

